updated IFacility, added a Facility class, refactored ChickenHouse
