+++ Interfaces/IFacility.cs
@@ -3,11 +3,11 @@ using Trestlebridge.Models.Animals;
 
 namespace Trestlebridge.Interfaces
 {
-    public interface IFacility<T>
+    public interface IFacility
     {
         double Capacity { get; }
 
-        void AddResource (T resource);
-        void AddResource (List<T> resources);
+        void AddResource (IResource resource);
+        void AddResource (List<IResource> resources);
     }
 }
